Handle SvelteKit breaking changes
- use shadow endpoints
- use URL instead of page
