+++ src/lib/util.ts
@@ -14,3 +14,17 @@ export function dedupe<Item>(arr: Item[], getId: (arg: Item) => any): Item[] {
 
 	return deduped;
 }
+
+// https://italonascimento.github.io/applying-a-timeout-to-your-promises/
+export function promiseTimeout(ms: number, promise) {
+	// Create a promise that rejects in <ms> milliseconds
+	let timeout = new Promise((_, reject) => {
+		let id = setTimeout(() => {
+			clearTimeout(id);
+			reject('Timed out in ' + ms + 'ms.');
+		}, ms);
+	});
+
+	// Returns a race between our timeout and the passed in promise
+	return Promise.race([promise, timeout]);
+}
