+++ svelte.config.js
@@ -8,8 +8,6 @@ const config = {
 	preprocess: preprocess(),
 
 	kit: {
-		// hydrate the <div id="svelte"> element in src/app.html
-		target: '#svelte',
 		adapter: netlify()
 	}
 };
